+++ packages/ledger-icrc/src/ledger.canister.ts
@@ -9,9 +9,10 @@ import type {
   Allowance,
   BlockIndex,
   GetBlocksResult,
-  _SERVICE as IcrcLedgerService,
-  Tokens,
   icrc21_consent_info,
+  _SERVICE as IcrcLedgerService,
+  StandardRecord,
+  Tokens,
 } from "../candid/icrc_ledger";
 import { idlFactory as certifiedIdlFactory } from "../candid/icrc_ledger.certified.idl";
 import { idlFactory } from "../candid/icrc_ledger.idl";
@@ -221,4 +222,14 @@ export class IcrcLedgerCanister extends Canister<IcrcLedgerService> {
 
     return response.Ok;
   };
+
+  /**
+   * Returns the list of standards this ledger supports.
+   *
+   * @link: https://github.com/dfinity/ICRC-1/blob/main/standards/ICRC-1/README.md#icrc1_supported_standards
+   *
+   * @returns {Promise<StandardRecord>} The list of standards.
+   */
+  getSupportedStandards = (params: QueryParams): Promise<StandardRecord[]> =>
+    this.caller(params).icrc1_supported_standards();
 }